fix: sidebar weird bug admin
