+++ react-app/src/view/UserAuth/RequireAuth.jsx
@@ -1,11 +1,27 @@
 import { Navigate, useLocation } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
 import { useEffect } from "react";
-import { setAxiosToken } from "../../axios";
-
-import { setToken } from "../../store/auth";
+import axiosInstance, { setAxiosToken } from "../../axios";
+import { base } from "../../api";
+import { setToken, signIn } from "../../store/auth";
 const RequireAuth = ({ children }) => {
   const dispatch = useDispatch();
+  const auth = useSelector((state) => state.auth.isAuthenticated);
+  let location = useLocation();
+
+  const getUser = async () => {
+    try {
+      const { data } = await axiosInstance({
+        method: "get",
+        url: base + "/auth/get-user-info",
+      });
+      dispatch(signIn(data.data.attributes));
+      // setAdmin(data);
+      console.log(data?.data?.attributes);
+    } catch (errro) {
+      console.log("not successful");
+    }
+  };
 
   useEffect(() => {
     const token = localStorage.getItem("token");
@@ -15,10 +31,9 @@ const RequireAuth = ({ children }) => {
     } else {
       setAxiosToken();
     }
-  }, []);
 
-  const auth = useSelector((state) => state.auth.isAuthenticated);
-  let location = useLocation();
+    if (auth) getUser();
+  }, []);
 
   if (auth) return children;
   else return <Navigate to="/login" state={{ from: location }} replace />;
